fix: extraInfo and avatar are not reflecting on UpdatePeer and AddPeer  (#601)
* fix: peers avatar and extraInfo not updating on redis

Everywhere we have type declaration and usage for avatar and extraInfo for peer. But not set on redis while creating peers.

* fix: SA5011 lint error on room_signal file

Before pinfo null check routines returning error sid and uid are dereferenced.

Moved the dereference after error handling for null check.

* fix: extraInfo and avatar are not reflecting on UpdatePeer and AddPeer

On UpdatePeer event and AddPeer  methods there is missing implementation for updating or adding avatar and extrInfo of a peer.